extarct the AnyRealmValue extention methods to dedicated file
